requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bson==0.5.8
dataclasses==0.6
Django==2.2.27
djangorestframework==3.13.1
djangotoolbox==1.8.0
dnspython==2.2.1
mongoengine==0.24.1
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.2.4
